rename analytics to amplitude
